+++ MSAL/src/native_auth/public/state_machine/delegate/SignInDelegates.swift
@@ -57,8 +57,9 @@ public protocol SignInStartDelegate {
 
     /// Notifies the delegate that a multi factor authentication (MFA) is required from the user to continue.
     /// - Note: If a flow requires this optional method and it is not implemented, then ``onSignInStartError(error:)`` will be called.
+    /// - Parameter authMethods: The list of available authentication methods for MFA.
     /// - Parameter newState: An object representing the new state of the flow with follow on methods.
-    @MainActor @objc optional func onSignInAwaitingMFA(newState: AwaitingMFAState)
+    @MainActor @objc optional func onSignInAwaitingMFA(authMethods: [MSALAuthMethod], newState: AwaitingMFAState)
 
     /// Notifies the delegate that the sign in operation completed successfully.
     /// - Parameter result: An object representing the signed in user account.
@@ -77,8 +78,9 @@ public protocol SignInPasswordRequiredDelegate {
 
     /// Notifies the delegate that a multi factor authentication (MFA) is required from the user to continue.
     /// - Note: If a flow requires this optional method and it is not implemented, then ``onSignInPasswordRequiredError(error:)`` will be called.
+    /// - Parameter authMethods: The list of available authentication methods for MFA.
     /// - Parameter newState: An object representing the new state of the flow with follow on methods.
-    @MainActor @objc optional func onSignInAwaitingMFA(newState: AwaitingMFAState)
+    @MainActor @objc optional func onSignInAwaitingMFA(authMethods: [MSALAuthMethod], newState: AwaitingMFAState)
 
     /// Notifies the delegate that a registration of a strong authentication method is required to continue.
     /// - Note: If a flow requires this optional method and it is not implemented, then ``onSignInStartError(error:)`` will be called.