+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [TBD]
+* Awaiting MFA Delegate now automatically returns the AuthMethods to be used when calling MFA Challenge #2687
+
 ## [2.4.0]
 * Prevent double conversion of MSALInternal error #2688